Eliminate negative indexing from ANY-WORD!
In R3-Alpha the payload structure of a word value suggested that words
were bound to "frames", and you could get at the frame (pair of series
of keys and vars) with VAL_WORD_FRAME(word).  The index in that frame
was retrieved with VAL_WORD_INDEX(word).

This was a fiction because the "frame" field, wasn't always a frame,
but was a "word series" for a function.  Then sometimes the index would
be negative to reflect that when you said VAL_WORD_FRAME you weren't
getting a frame, but one of these series instead.  The variables would
be found by walking the stack, and then picking from the stack level
using the absolute value of the index.  Further confusion was added
by 0 meaning "self".

To try and make this more clear, Ren-C altered the binding of the
words from VAL_WORD_FRAME to "VAL_WORD_TARGET" and discerned in the
system when a REBFRM was being used vs. a REBARR that was a function's
"paramlist".

With the word "CONTEXT" now the superclass of OBJECT! and also the
superclass o a more persistent form of function instance invocation
known as the retaken term for FRAME!, this does an adjustment.  The
ANY-WORD! has VAL_WORD_CONTEXT which always returns a REBCON*.  Then
it is assumed that an ANY-CONTEXT! can be tested for whether it is
a FRAME! or not...removing the need to mark this with a negative
number.

The current stand-in for a "FRAME! context" is still the relatively
weak use of a function's paramlist, which does not indicate each
invocation uniquely.  However, it can be discerned from other kinds of
contexts already...because the slot 0 value in the paramlist is an
ANY-FUNCTION!.  This is good enough to do the test that makes the
negative indices unnecessary.