Add FRAME! datatype, solidify ANY-CONTEXT!
This adds the pending FRAME! datatype for tracking a specific instance
of a function call (with no implementation yet).  The "context" label
for the superclass (that used to be ANY-OBJECT!) is propagated more
thoroughly into the type table and the naming of the dispatch
functions that are referred to or generated.

This also corrects tab alignment in the types table.