Check ANY_FUNC before checking EXT_FUNC bit
The <infix> idea is slated to be killed to bring back a separate
datatype (likely INFIX!, synonym OP!)...as it represents a unique
interface contract (and as such, doesn't belong in whatever the
ANY-FUNCTION! category is called).

That aside, there is a check for a type-specific "infix" bit that is
only applicable to functions, being made in the Do_Core loop before
testing to ensure that it's a function.  This risks another type that
uses the same bit for another reason being interpreted as an infix
function.

Fixed by adding a function check, but this is (again) temporary.