Audit VAL_RESET_HEADER vs. VAL_SET_TYPE
In R3-Alpha there was VAL_SET and SET_TYPE, both of which took a value
and set it to a type.  The difference was that VAL_SET cleared the
header bits out, while SET_TYPE only affected the bits of the type
that were relevant to the header.

A recent change renamed these to VAL_RESET_HEADER and VAL_SET_TYPE.
This is an audit which looks at usages to see which was actually
intended.  Several cases of just setting the type bits were in the
code...which is bad, because even if a type has no per-type flags
the value bits like THROWN and whether a newline should be put out
in the molding could wind up corrupt.

In addition, this renames VAL_SET_TYPE to VAL_SET_TYPE_BITS to further
suggest that a "whole thing" is not being set, only a partial.  This
would hopefully cue the notion that the value must already be
initialized in some way.  A more rigorous debugging check could be
added which discerned a value that had its header set from one that
had not, but that may be overkill considering how few calls there are.