Move bound state into word flag (vs payload)
Historically the unbound state of a word has been conveyed by the
pointer to the context (frame, etc) being null.  In this state the
index is meaningless; Ren-C debug builds forced it to a special value
that could be checked, but it was essentially unused.

Having the indicator of whether a word is bound or not inside the
payload of the REBVAL enforces some amount of common structure between
the payloads of bound and unbound words.  The thinking for how to
successfully bind uniquely to each function instantiation requires
more space opened up in the payload of a bound word to account for a
differentiating factor, so more fluidity is necessary.

This pushes the bound or unbound state into a flag in the header,
making it possible for bound words to be structurally completely
different in their payloads from unbound words.  This is a forward-
looking modification to a possible unification of ANY-STRING! with
ANY-WORD!, in which the unbound state of words are interned strings.