More series flags, add SER_PARAMLIST
This expands the capacity for the number of series flags to more than 8
by moving the bias information into the dynamic portion of the series
content (which will be used to hold series data in small series,
when that feature is implemented).

The specific motivator for doing it at this time was the need for a
flag to help distinguish a FRAME! from a non-FRAME! context, without
needing to look into the array data of the pointer.  This is required
due to the order of creation and asserts.  It also helps performance,
but the ultimate layout when FRAME! is fully implemented will likely
be very different so that is probably irrelevant.