Context macro cleanup, INIT_CONTEXT_KEYLIST
Small maintenance on some context-oriented macros (and series, array
a bit) which had gotten out of sync with the lettering as the typenames
have been changed over time.

Includes a separation of fetching the context keylist from initializing
it, where the initialization does not require the context to be in
a finished/assertable state.