Lock functions, legacy function-bodies-mutable switch
With specific binding, all function instances get their own "view" of a
common body of relatively-bound code.  This means their word bindings
are specific to their arguments and locals, and not confused with the
recursions.  Given that the functions are sharing the same body, it makes
it further clear that it should not be the case that functions modify
their bodies...as they cannot do so with their arguments in a way that
would relativize to the other calls.

This goes along with a general push that more code is likely to be read
only by default, e.g. that coming back from LOAD, and that rebound views
of code will be cheap to make without needing to copy.  Those who wish
to copy can do so.

So in this case, if one wants a mutable function body it's easy enough
to do so...by making the body a DO of a mutable block that one gets from
somewhere (for instance, just by copying.)  This is a mutable function:

    foo: function [x] [
        do code: copy/deep [
            block: []
            append block x
            probe block
        ]
    ]

If the pattern is common enough, it could be a feature of function, e.g.
<self-modifying>.