Enhance IS_END check to fail on unsafe trash
IS_END was previously willing to say that a "trash" slot was not an
end, even though various other tests on it would fail (e.g. VAL_TYPE)

Abstractly speaking, an IS_END check can't guarantee that something
with the trash bit pattern in the type slot is not an END, because it
might be a pointer "doing double-duty" as an END.  However, it is
possible that if the WRITABLE bit is set to put a type in the type
bits that isn't trash when writing ends.  This way END can check those
for trash and assert if it is.

Safe trash must be excluded from the check, and the debug value of
REB_MAX will not be usable if there wind up being 63 total Rebol base
types.  In that case, a semi-obscure non REB_TRASH value could be
used.  But until that point, REB_MAX is a better choice to help hint
that it's an END marker vs. a trash.