Reverse sense of TRAP in Get_Var, documentation
The parameter `trap` to Get_Var_Core suggested whether errors should
be raised or NULL returned.  Considering the functionality of the TRAP
native, the sense of the flag was inverted.

If a function receives a flag that says "trap" that should suggest that
an error *not* be bubbled up, rather be "trapped" by the call.  It's
a little bit confusing because it's the call that's generating the
error in the first place, but from the caller's point of view who is
passing in the trap flag it makes sense to not have errors bubbling
out if that flag is TRUE.

Also adds a bit of documentation to the header file.