Add "get fallthrough" to legacy GET primitive
The GET native in R3-Alpha would pass through any type it didn't know
how to GET.  (So `get $10` was $10, and `get <fork>` was <fork>`).
Ren-C and Red are more constrained and give errors on things that
it isn't able to use as a specification of a variable to read from.

R3-Gui apparently had a contingency on the old behavior:

    if error? err: try [
            dial: delect guie/dialect orig-dial: dial opts
        ] [
            either all [
                word? act: dial/1
                block? body: get dial/2
            ] [
            ] [
            ]
        ]

This adds a legacy proxy GET to allow the old behavior.