Make GET of ANY-CONTEXT! work with hidden fields
A feature in R3-Alpha allows one to get all the value contents of an
object (no keys) via GET of that object.  It was not able to do this
if there were any fields that were marked hidden, in which case it
was not considered a "public object", even if other fields were not
being hidden.

With Ren-C switching SELF to be an ordinarily bound key in the object,
it was marked hidden...making contexts with hidden fields the norm.
This adds the functionality of letting GET fetch just the public values
in the context so that GET of an ANY-CONTEXT! will work again.  It
also gets rid of the "public object" term and Assert_Public_Object.

The larger question of whether this should be a feature of GET or
another reflector (VARS-OF ?) remains open.