Running `<r3-legacy>` multiple times => noop
If r3-legacy is allowed to run multiple times, that means that the
mode switching code (and the support functions) need to be able to run
in a lowest-common-denominator implementation of Rebol between what
legacy mode supports and how Ren-C is configured out of the box.
As that is not a desirable effort to be engaged in, this makes any
subsequent runs of r3-legacy after the first a noop.