Mold stack with Push/Pop/Drop abstraction
The "mold buffer" is a longstanding concept of a large per-thread
series of REBUNI characters (2 bytes each).  Because requests to
generate unicode strings are frequent and allocating and freeing series
has overhead, the idea was that this buffer would be reused.

As the name suggests, one of the big re-users of this buffer is MOLD.
But it wasn't the only client, as several other operations that did not
think the mold buffer was in use would use it as well.

This had a downside--in that if one ever tried to do a debug print that
wound up molding or forming a value (as many would like to do) then if
that were called in one of these routines that also used the mold buffer,
the work would be destroyed.

Other problems came up, such as that a routine like Form_Reduce
which wished to reduce items and "mold as it went" could not
"evaluate=>add to mold buffer=>evaluate=>add..." because the evaluation
might call something that would use the buffer and overwrite it before
Form_Reduce was finished.  Hence it had to save up all the values from
all the evaluations and then begin the mold, uninterrupted.

This implements the mold stack as an abstraction layer that anyone who wants
to use the mold buffer has to go through the REB_MOLD "api" and can
push content to the buffer at a marked start position.  It can pop the data
off as either a unicode string or as a UTF8 binary, or it can examine the
unicode characters in the buffer directly for whatever purpose it has and
then "drop" it to the marked position at the start of that stack level's mold

The layering is imperfect, and the naming of everything being "molding"
is inaccurate (but no less accurate than using BUF_MOLD).  Several
pieces were changed to support the uniform interface, such as letting
the debug format routines use unicode and expand the buffer (vs. a
cut off fixed size as before).  Specialized routines for printing that
would implement a limit have been updated to a setting for mold
itself of MOPT_LIMIT, as well as the addition of the ability to reserve
an amount of space with MOPT_RESERVE if the mold buffer is to
be worked with directly.