TRUE? FALSE? QUOTE? SOMETHING? NOTHING? natives
This adds natives for these basic routines.  The implementations are
not currently frameless, but could be made so.

Also adds /BODY refinement to NATIVE for supplying an "equivalent body"
of Rebol code that would behave the same as the native.  It should
be observed that if one actually ran the alternative implementation,
in the general case it could only be one alternative implementation at
a time--because if natives were implemented in terms of themselves
without any C behind them they could recurse infinitely.