Remove C-string label debug tool
As a debugging tool, an option to save a C string pointer inside of a
series node existed in R3-Alpha.  It wasn't intended to be user-facing,
as some kind of generic "label a series with another series" feature.

Superior ways of labeling the origin of series come from Panic_Series
under Valgrind and address sanitizer, which effectively labels *every*
series with its point of allocation...also without requiring a build
switch.

Series labeling had been controlled by a build switch that had not been
turned on since the open sourcing of R3-Alpha, and considering the
large amount of series macros and code being shuffled for NewFunction
the clutter of the labels are removed.