Copy_String_Slimming, use in mold stack
The Mold Stack unified several different string copying routines and
used Copy_Sequence_Len as the basis for extracting the string via
Pop_Mold_Stack.  It left the idea of figuring out whether the produced
string could be made a REBYTE string as a "TBD".

It turns out this functionality existed already in Copy_String.  This
exposes that better by renaming that to Copy_String_Slimming and uses
the routine in the implementation.

(Theoretically the mold stack could track the data as it was added and
note if any level would need more than a byte sized series.)