Don't reuse REBUNI-wide series for UTF-8
There is a somewhat questionable routine for making temporary series
on an "as-needed" basis for UTF-8 encoding.  In it, it was willing to
reuse a 0 length REBUNI-wide string, and pass it on to code that would
try using it as a UTF-8 buffer.  Due to a lack of size checking or
asserts, these routines would use the string assuming it was byte
sized.

To appease upcoming tightened asserts (and avoid the bugs that not
appeasing them would likely entail), this doesn't consider the 0 case
as an exception.  The real answer to this routine is likely not to
burden the GC with these temporaries in the first place, as the
lifetime must be known (else using the temporary would be unsafe).

It also protects the series in the case of making a copy, in order to
hint that *neither branch* should be modifying the resulting series.