Integrate cast+typecheck to series data access
When a Rebol series is created, it is created with an element size
(a.k.a. the "width").  Since this is C there is no templated way to
store the element type.  So there were macros that would get a
pointer at a position in a series.  These were called, and then cast
by the caller.

This integrates the cast into the macro, and adds an assert.  So where
before one would write:

    Foo *foo = cast(Foo*, SERIES_DATA(series_of_foo));

One now writes:

    Foo *foo = SERIES_HEAD(Foo, series_of_foo);

(The reason for changing it to SERIES_HEAD is to avoid there being two
names for the same routine, as the XXX_DATA and XXX_HEAD were
the same thing before.)

Inside the macro, it is asserted that the size of the type passed in
matches the series width, and a pointer to the passed in type is the
type of the return.

This enforcement runs up against those cases that work with series that
could be multiple different sizes.  For those, SERIES_DATA_RAW offers
the REBYTE* of the data without any checking, as a synonym of the
old form of SERIES_DATA, and as a stronger distinction from HEAD.

A similar change is included for SERIES_AT and SERIES_AT_RAW.

Two circumstances required modification: the random seed of a string
(was only written to handle byte sized series, needed to be changed).
Also the LIB function lookup was assuming the string it got in was
byte sized, when it could also be a REBUNI-sized string.

This includes some documentation notes about the UNI and BIN series,
and a general alignment of those macros as well.