Add COLLECT-WITH mezzanine to specify "keeper"
Motivated by a question about the limits of COLLECT by Darek Nędza,
this reframes COLLECT in terms of a basic building block COLLECT-WITH,
so `collect [...]` is the same as `collect-with keep [...]`

In addition to letting you choose a name besides KEEP for the word
that is assigned the keep function, it allows you to preserve a
*binding* by passing in a LIT-WORD!.  Doing so indicates there is
no need to deep copy or rebind the collect body inside COLLECT-WITH,
it needs only to set the variable to the collecting function.

    >> my-keep: none

    >> foo: func [] [
        my-keep "this will get collected"
    ]

    >> collect-with 'my-keep [foo]
    == ["this will get collected"]

Using an ordinary word will go through the function-based binding and
creation of a new variable (as with the old collect).