SER_XXX => OPT_SER_XXX, documentation
Motivation for the name change is that there are going to be two kinds
of series pointers to talk about: the double-wide kind of series
pointer that carries around a FRAME! with it, and the lower-level and
ordinary REBSER* as we know it.  To lessen the confusion, routines
that operate on the lower-level construct will operate on a SER while
the higher-level constructs will operate on an HSERIES.

Given that there will be routines like SER_LEN that stand in contrast
with SERIES_LEN, having something named SER_LOCKED would just be
confusing.  So this sticks the OPT_ in front.

As long as the change was being made, added more explanation on the
series flags for good measure.