Change `context` field of Any_Word to union
A historically confusing "lie" was the idea that the pointer for the
binding in a word always pointed to the same kind of object--when
there were two very different implementations.  With the plans set
forth for FRAME! and NewFunction, it is now more important to discern
a relatively bound word from one bound to a "normal" context...and
there is another (and probably two more) categories of binding to
deal with.  (Extracted relative references, to frames and then most
likely to per-frame objects, for a CLOSURE!-like expansion of behavior)

This splits out the context information into two distinct members of
a union, with which is in effect signaled by header bits.