GROUP! as new default term for PAREN!
This changes the convention for Ren-C to use the name GROUP internally
such that the macros are IS_GROUP() instead of IS_PAREN(), REB_GROUP
instead of REB_PAREN, etc.  PAREN!, PAREN?, and TO-PAREN are kept
as synonyms in the mezzanine, and a legacy switch to get TO-WORD
of the datatype to say "PAREN!" is also included.

(Long term the hope is to deprecate TO-WORD on datatypes, to eliminate
their keyword form and use exemplars, along the lines of #[()] for
a GROUP!/PAREN! so the word doesn't matter.)

Making this change and type renames less of an issue in Ren-C is the
new behavior of switch.  Previous concern over what name a type had
was often driven by switch statements, where a switch on TYPE?/WORD
could only have one answer regardless of what the alias one used was.
With TYPE-OF no longer offering a word and switch being able to
fetch values with get-words, the answer is better:

    switch/all type-of quote (a b c) [
        :paren! [print "This should work fine."]
        :group! [print "So will this."]

        paren! group! [
            print "**Neither** of these would work."
            print "You must compare to a datatype!"
        ]
    ]