Enable memory leak reporting in debug build
When the "zero leak Rebol" was first put forth, it was possibly too
disruptive to alert users to leaks, even if it was just the debug
build.  Given that no leaking has been holding stable, this now makes
the leaking "noisy" should it ever happen.

Rather than assert, it prints the amount of memory leaked--making it
easier to get a good report from Valgrind.