Move dumping code under #if !defined(NDEBUG)
There is a lot of R3-Alpha dumping code that is not being used or
maintained.  It has been kept around in case it would be useful if
there were some problem with the memory pools, and sometimes used
as a reference.

However R3-Alpha did not have a debug/release build separation.  This
pushes some dump routines into the debug build only.