Ensure release build doesn't use <stdio.h>
The Rebol Core was intended to not have dependencies on printf() or
other standard library I/O.  Specifically, the evaluator and its
services were to use the host API.

Although the Rebol "formatted print" is more powerful in a generic
sense, it means that even the simplest debug output goes through a
complex set of translations...possibly unicode to UTF8, and then
through the host kit.  This may be the very code one is trying to
debug, so using Debug_Fmt can be a problem.

This splits it so that the debug build does a blanket include of
`<stdio.h>`.  However, the release build has some garbage definitions
which will prevent compiling with stdio.h.

Some third party files included stdio.h, but seemingly none of them
need it in the Rebol core--as commenting out the lines gave a working
release build.