Use REBUPT for alignment in series
The desire to only have REBVALs and pointers on pointer-aligned
locations had some conditional code on __LP64__/__LLP64__.  This is
more purposefully accomplished with REBUPT (Rebol equivalent of
uintptr_t), with only a single struct field to document.

Also added a paranoid alignment check, which was run and verified,
then commented out and left as a reminder to make a "extra checks"
debug build that does costly things like that.