Tidy %u-parse.c to 80 column limit
The parse implementation had many lines that were off the side of the
screen, that have gradually been reigned in over time as part of
encroaching cleanups.  This finalizes it by bringing the whole file
into the 80 column limit.

It also takes a cue from Do_Core where the state variable that is
generally implicit (the parse state) is renamed from `parse` to just
`p`.  By abbreviating it, the need to create local variables which
alias the parse state is removed (as `p->series` is nearly as short
as just `series` and more explicitly locates it as part of the
parse state vs. a different local variable).