Add a series flag OPT_SER_ACCESSIBLE
When the series has an external data storage, the life cycle of the data
storage is not managed by the core. The core needs to know when the data
can not be accessed any more, or the core could crash.
`OPT_SER_ACCESSIBLE` is for this purpose.

This commit just implements the basis for notifying the core that the
data is no longer accessible. A following commit will apply this
machnism to struct! where EXTERNAL data storage is used.