const-correct Compare Type (CT_ functions)
The compare type functions take pointers to REBVALs they should not be
modifying in the course of comparison, so they should be const.

(Note that there is a modifying Compare_Type root function that is
destructive to REBVALs that it assumes you don't mind ruining, which
does this modification in order to reuse conversion work to convert
one of the two types you give it to see if it matches after conversion.
Rightly or wrongly for its existence, that is properly non-const,
given how it works.)