Revert "Check for END in IS_* macros"
This reverts commit 7a4607733c65e204d70d32db28dcfab7d6a1ce63.

Per discussion with @ShixinZeng, the choice to cause a failure in the
macros on allowing an END marker when testing via ordinary type checks
was intentional.  It is a difference from the macro behavior in
R3-Alpha, but has largely drawn attention to logic errors in code that
didn't realize it was not processing a value.