Move /RETURN on BREAK handling to mezzanine
This makes LIB/BREAK a mezzanine that wraps the native break in order
to provide the /RETURN refinement.  It also rewords the help strings
on BREAK and CONTINUE to better convey what the behavior of /WITH
really is--as it has been formulated as a refinement (no pun intended)
of /RETURN's behavior in R3-Alpha.