Optimize/localize definitional return arg filling
As an initial cut at helping the performance of definitional return,
the search for a definitional return word was unified with parameter
filling...as the RETURN: parameter could be theoretically anywhere
with generator-added locals after it (possibly even at the start of
the parameter list).  This meant the list could be walked once.

(Note: theoretically the list could be walked once, during invocation,
but mechanically it is set to unsets before the work is done.)

Unfortunately, folding the testing and filling in with argument
fulfillment meant that more than one place in the code needed to do
it (Apply_Block, Do_Core, Redo_Function).  These routines were already
complicated, and were out of sync with each other...with bugs like
"Definitional return works if a closure is called from user code but
not if called from the C code as a system helper function" (which
unlike other such bugs remained undetected for some time, since there
were no system-helper closures).

With the impending commit of PROC adding LEAVE as another definitional
case that would have to be handled in multiple places again, a change
was needed to localize the code.  This makes it possible to localize
and avoid a third walk to do it by forcing the definitional return word
to the end of parameter lists that have them.  The point of dispatch of
FUNCTION! and CLOSURE! can handle it in one place.

Documentation and some related issues in the neighborhood of
modification were also cleaned up.