Fix assert/problem with CLOSURE!+THROW
Invoking a CLOSURE! with an argument that threw (`some-closure break`)
raised an issue about responsibility of management/freeing of arrays
used in argument gathering when the invocation did not get a chance
to run.

This fixes the issue, but also adds some notes on the general pattern
of changes and improvements pending in the related area of binding
and contexts.  It also fixes an over-aggressive assert in trying to
provide an early-warning about arrays handing trash to the GC.