Line break correction in definitional body
For a currently unknown reason, the line break from the source of the
function body template is not being carried through to the copy used
when making an instance.  This sets it, though a note is left to look
into why it wasn't preserved in the first place.

*(A casual guess that line compression during the boot code had taken
it out was not the case--it seems to be there.)*