Experminental EVAL/ONLY, subvert argument eval
EVAL gives special access to "bounce" the evaluator, giving its
argument inline access to subsequent parameters.  It replaces DO of a
FUNCTION! but is more general:

   >> eval (quote x:) 1 + 2
   == 3
   >> print x
   3

(See notes: https://trello.com/c/YMAb89dv )

This enhances EVAL with an /ONLY refinement, somewhat in the spirit of
APPLY/ONLY.  This allows one to slip the ensuing values directly and
bypass any evaluation, regardless of quoted or unquoted status:

    >> bar: func [a b :c] [print [a b c]]
    >> eval/only :bar 1 + 2
    1 + 2

The implementation of this feature is a further step toward making
Do_Core() the sole piece of code that is responsible for building
call frames in the codebase (such as doing complete refinement and
error checking on calls to Rebol functions from C, through a single
optimized path).

It is designed for experimenting with that mode, as well as being a
potentially useful feature in its own right.