Case-sensitive maps
Now maps are case-sensitive when setting values
but case-insensitive when finding.
Case-sensitive search is done with select/case.

m: map ["a" "a" "A" "A"] --> make map! ["a" "a" "A" "A"]
m/("A") --> "a"
select m "A" --> "a"
select/case m "A" --> "A"
(same for words; but chars are always case-sensitive)