Remove incorrect assertion in Drop_Mold
An assertion regarding was copy-pasted from the Pop_Mold code to the
Drop_Mold code regarding termination, when not all in-progress molds
terminate.  As diagnosed by @giuliolunati:

    ASSERT_SERIES_TERM(mold->series) in Drop_Mold_Core seems wrong:
    if pushed data are to be discarded, an unterminated mold->series
    may be right.   Indeed that happens when Scan_Item_Push_Mold
    returns NULL/ZERO.

This adds a No-Op NOTE_SERIES_MAYBE_TERM() in order to point out the
tolerance of an incomplete state.  This note can be used in other such
situations and searched for, in case a more consistent policy on
termination enforcement needs help being sized up...and to send a
signal to surrounding code to be careful and not use a routine on the
series requiring termination.