Change termination trick to not break aliasing
(Hopeful fix for #208, #211)

The idea of using the lowest bit in the header of a REBVAL to indicate
termination, and making it the size of a REBUPT (Rebol's uintptr_t
abstraction), led to the idea of allowing pointers in data structures
do this termination job as they were a sunk cost--and this would add
little complexity to those structures.

Though a bit sketchy in principle, suggestion that it worked on most
all platforms made it seem like something that could be used and
then disabled on those rare architectures upon which it would not work
in practice:

    http://stackoverflow.com/questions/19991506

But in addition to being overall a questionable idea to allow *any*
builds purposefully break undefined behavior, reports were submitted
of conventional platforms (Windows7 Intel w/MinGW, Linux 64-Bit w/gcc)
having trouble.

The existing deployment of the technique was only a test before a
larger application.  This rethinks it in a way that is believed to be
in line with the standard requirements of the compiler, and not
invoke undefined behavior.

The change is to not alias the REBUPT* to pointers to pointer-sized
things.  (That pointer cast is not legal, while -value- casting a
uintptr_t to and from a pointer is...these are different things.)
Instead the header bits of REBUPT are only interpreted out of
structures which layout a REBUPT at the tail position of the values
they wish to bound.