Dividing money! amount again returns money!, preserving precision
This reverts a3f249c7f30779563bdc761e0f2ee3a044c63b1a.

money! is a datatype designed to be particularly suitable (precise) for
typical monetary calculations. It is implemeted using an unnormalized
_decimal_ floating ploint format (in contrast to, say, the currently
unfortunately named decimal!, which uses IEEE-754 double precision
_binary_ floating point).

To not loose this desired precision, it is absolutely required that
results from calculating with money! operands are also given as money!.

Further, in the syntax used for money, values such as `$123.45` are
unit-less, i.e. _not_ dimensional. The unitary counterpart would be the
Rebol 2-like form `USD$123.45`, which is, however, not implemented in
Rebol 3 at this point.