Add EXT_TYPESET_EVALUATE to routine args
Every PARAM of a function should have some typeset bits set to say how
the evaluator should handle its arguments.  There was some leeway
falling through to a default of not forcing ROUTINE! to say and just
assuming evaluate, but with more asserts (as in tracing) enforcing
otherwise it was best to just find the place to add the bit and add it.