Drop SET? as "ANY-VALUE?" operation pending review
With refinement arguments being unset by default, adapting bits of code
that had assumed they were NONE! required rewriting or a way to check
for a set state.  Rather than NOT UNSET? the operation SET? was
added--with rationale that SET! would not be a data type (given MAP!
and other concerns) and that this was a common test.

A realization in chat while trying to find a more communicative name
for VALUE? led to this possibility:

    >> set 'x 10
    >> set? 'x
    == true

    >> unset 'x
    >> unset? 'x
    == true

It paints a rather clear picture that SET? and UNSET? might be seen as
operations on variables as opposed to values.  (Bound words being the
closest thing Rebol has to representing "variables", as they are the
access points for both reads and writes.)  However it also raises a
number of questions.

So whether that decision goes forward or not, it made it seem hasty at
this point in time to pre-decide on SET? as the opposite of UNSET?
for values.  The ANY-VALUE! category typeset is already used in
function specs and considered "anything but UNSET!" (pending upcoming
change, it's still a synonym for ANY-TYPE! until #opt vs <opt> is
decided for function specs.)

Given that ANY-VALUE? is a safe bet, and may ultimately evolve into
simply VALUE? (if legacy uses can be corraled in relevant code or
given compatibility shims until they can be adapted), it replaces
SET? in this commit.  There are few uses so best to back them out
now that the question exists before they become widespread.