Add CHECK-SET routine per suggestion from @earl
The problem which allowing `x: ()` to not be an error was designed to
address can be handled in a less system-breaking way with a primitive
suggested by @earl, which is CHECK-SET...an arity-2 function that
soft-quotes its first parameter, which is constrained to be either a
set-word or a set-path:

     if check-set var: thing-that-may-be-unset [
         ...
     ]

Idiomatically this comes across reasonably well that var: is being
set by this code.  CHECK-SET does not come off most importantly
as a query due to the lack of a question mark, rather as some form
of setting that involves a check.  Even if glossed as CHECK-SET? it
still fits well, as it seems an assignment has happened whether you
realize var: is a parameter to check-set or not.

The promise of this option outweighs the loss in locality of error
delivery given by allowing `x: ()`--hence that will be backed out
shortly.