SET/ANY and GET/ANY => SET/OPT and GET/OPT
This was a long-pending modification being motivated by taking another
step toward straightening out unsets.  The /ANY refinement has no
bearing on optional setting (and ANY has another much more common
meaning) while the /OPT refinement corresponds directly to the idea
of a value which can be unset as well as carry a range of other values.

GET already had a legacy mezzanine to support its fallthrough of
being able to get values like `get 3`, an ability that Red also did not
carry forward.  This moved SET out as well--which is forward-thinking
based on the idea that the current interface of SET for setting objects
positionally breaks the rule that objects aren't supposed to be
accessed by position.  Hence rather than pollute the core with a
deprecated refinement this moves the refinement out.

Having the versions of SET and GET be in user code will slow them
down some, which is unfortunate...but it's a temporary state as the
/ANY versions will be legacy mode after a sufficiently long time is
given with /OPT being offered for code to adapt.  There are also some
pending designs for speeding up such wrappers in a way that might not
require new bodies just to rename refinements (though as mentioned,
that's not necessarily all these particular wrappers will do.)