Differentiate OBJECT! and ARRAY access in protect
In R3-Alpha, it was sometimes the case that the "frame" series of an
object was accessed via VAL_SERIES instead of what was then VAL_FRAME
(now VAL_CONTEXT_VARLIST).  Most of these have been vetted from the
system to go through the accessor appropriate for the type, but this
case in PROTECT had not been changed.

Caught by a runtime check in the debug build.