Format non-series memory for stack REBVAL
The new strategy for "sunk cost" terminators in Rebol ARRAY is helped
to be more robust by testing a "writability" bit to prevent code that
manipulates the arrays from assuming ownership of the value that is
doing double-duty as a terminator.  This writability bit is taken care
of automatically when array series are created, but C stack values
have to be manually initialized.  This was a missing initialization.