Hopeful cleaner solve to aliasing problem
This is a hopefully-now-standard workaround for the bug #208 #211,
which did indeed turn out to be undefined behavior from aliasing.

It *appears* that by taking the address of a struct field of a certain
type and having that address in a pointer, then writing to the data via
that pointer, will enforce coherence with other locations known to be
of that type.

This suggests that different semantics apply to:

     Foo *foo = &bar.foo;
     *foo = some-foo;

vs.

     bar.foo = some-foo;

Where seemingly, the write to the latter only forces coherence with
other foos living in Bars, not all Foos system-wide.