no refinements on DO* helper, END_VALUE varargs
DO* is a system function which is called only from the C code.  It
depended on positional calling of refinements, an ability which is
being removed from Ren-C.  There will be alternative ways to invoke
refinements at the low and high level, but they will not be positional
in nature...hence refinements can be added, removed, and reordered
without concern--even from system functions.

This switches DO* to use UNSET! as the out of band value and removes
the refinements, turning it into a fixed arity function.

Also, this changes the terminal used to signal the end of varargs
invocations with REBVAL* to END_VALUE, a pointer to a global end
marker.  While NULL might seem briefer and more convenient, it is
actually incorrect, because it is defined as 0 and in a varargs
context you would have to say (REBVAL*)NULL in order for it to know
that it was a pointer type and not an integer at the receiving side.

https://en.wikipedia.org/wiki/Stdarg.h#Type_safety

Incorrectness of NULL pointed out by @Morwenn via a report found in
CppCheck.