HALT as Ctrl-C instead of BREAKPOINT
In order to demonstrate the debugger and begin testing it, the Ctrl-C
signal was taken temporarily for the purpose.  Under real usage this
is not desirable as a default, but rather only under a "debug mode".

Though there is a "debug mode" already, it has not been planned out
the point of decisions like where the flag would be and how the place
where Ctrl-C is received would be wired up to the decision of which
action to take.  This reverts the flag to the halting default until
those decisions are made, which will not be the next item scheduled.