Use temp buffer and not UNI_BUF in Read_Dir
Generally speaking, Rebol's design surrounded reusing a lot of "hot"
temporary buffers rather than paying for new allocations.  The
technique has been made nicer with Mold Stack so that each new
re-user of the buffer "pushes" their buffer need and pops it when
it is done.

There was one outstanding use of UNI_BUF that had been aliased through
a macro for "BUF_OS_STR".  This was in Read_Dir, and the upshot was
that on windows the Read_Dir function would erase the mold stack
in the now-cooperative protocol...then leave junk in it.  This was
causing an assert in the windows build, where the wide character
similarity of REBUNI to Windows OS API character strings was being
exploited.

Rather than make Read_Dir speak the Mold Stack protocol, this just
allocates a temporary buffer...as the time spent reading the directory
will easily eclipse any concern about a small buffer being used out
of the memory pool.